changes to get bookmarks in fragment
