+++ app/src/main/java/org/openstaxcollege/android/activity/ViewBookmarksActivity.java
@@ -11,16 +11,20 @@ import java.util.Collections;
 
 import android.app.ActionBar;
 import android.app.Activity;
+import android.app.FragmentTransaction;
 import android.support.v7.widget.DefaultItemAnimator;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.support.v7.widget.helper.ItemTouchHelper;
+import android.text.Html;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.Window;
 import org.openstaxcollege.android.R;
 import org.openstaxcollege.android.adapters.BookmarkRecyclerViewAdapter;
 import org.openstaxcollege.android.beans.Content;
+import org.openstaxcollege.android.fragment.BookmarkFragment;
+import org.openstaxcollege.android.fragment.GridFragment;
 import org.openstaxcollege.android.handlers.MenuHandler;
 import org.openstaxcollege.android.providers.Bookmarks;
 import org.openstaxcollege.android.providers.utils.DBUtils;
@@ -39,27 +43,27 @@ import co.paulburke.android.itemtouchhelperdemo.helper.SimpleItemTouchHelperCall
  * Used to display cardview of saved Favorites
  *
  */
-public class ViewBookmarksActivity extends Activity implements OnStartDragListener
+public class ViewBookmarksActivity extends Activity
 {
-    /** Adaptor for Lens list display */
-    BookmarkRecyclerViewAdapter adapter;
-    RecyclerView recyclerView;
-    /** list of lenses as Content objects */ 
-    ArrayList<Content> content;
-    
-    /**handler */
-    final private Handler handler = new Handler();
-    
-    /** Inner class for completing load work */
-    private Runnable finishedLoadingListTask = new Runnable() 
-    {
-        public void run() 
-        {
-          finishedLoadingList();
-        }
-      };
-
-    private ItemTouchHelper itemTouchHelper;
+//    /** Adaptor for card display */
+//    BookmarkRecyclerViewAdapter adapter;
+//    RecyclerView recyclerView;
+//    /** list of bookmarks as Content objects */
+//    ArrayList<Content> content;
+//
+//    /**handler */
+//    final private Handler handler = new Handler();
+//
+//    /** Inner class for completing load work */
+//    private Runnable finishedLoadingListTask = new Runnable()
+//    {
+//        public void run()
+//        {
+//          finishedLoadingList();
+//        }
+//      };
+//
+//    private ItemTouchHelper itemTouchHelper;
       
       /* (non-Javadoc)
        * @see android.app.Activity#onCreate(android.os.Bundle)
@@ -69,36 +73,48 @@ public class ViewBookmarksActivity extends Activity implements OnStartDragListen
       public void onCreate(Bundle savedInstanceState) 
       {
           super.onCreate(savedInstanceState);
-          requestWindowFeature(Window.FEATURE_INDETERMINATE_PROGRESS);
-          setContentView(R.layout.card_view);
-
+          setContentView(R.layout.bookmark_layout);
           ActionBar aBar = getActionBar();
           aBar.setTitle(getString(R.string.title_favs));
-          aBar.setDisplayHomeAsUpEnabled(true);
-          setProgressBarIndeterminateVisibility(true);
-          recyclerView = (RecyclerView)findViewById(R.id.recycler_view);
-          recyclerView.setLayoutManager(new LinearLayoutManager(this));
-          recyclerView.setItemAnimator(new DefaultItemAnimator());
 
-
-          //get already retrieved feed and reuse if it is there
-          content = (ArrayList<Content>)getLastNonConfigurationInstance();
-          if(content == null)
-          {
-              //no previous data, so database must be read
-              readDB();
-          }
-          else
-          {
-              //reuse existing feed data
-              adapter = new BookmarkRecyclerViewAdapter(content, R.layout.card_row, this);
-              recyclerView.setAdapter(adapter);
-              ItemTouchHelper.Callback callback = new SimpleItemTouchHelperCallback(adapter);
-              itemTouchHelper = new ItemTouchHelper(callback);
-              itemTouchHelper.attachToRecyclerView(recyclerView);
-              setProgressBarIndeterminateVisibility(false);
-             
-          }
+          aBar.setDisplayHomeAsUpEnabled(false);
+          //if(getFragmentManager().findFragmentByTag(GRIDFRAG) == null)
+          //{
+              FragmentTransaction transaction = getFragmentManager().beginTransaction();
+              BookmarkFragment fragment = new BookmarkFragment();
+              transaction.add(R.id.container, fragment);
+              transaction.commit();
+          //}
+//          requestWindowFeature(Window.FEATURE_INDETERMINATE_PROGRESS);
+//          setContentView(R.layout.card_view);
+//
+//          ActionBar aBar = getActionBar();
+//          aBar.setTitle(getString(R.string.title_favs));
+//          aBar.setDisplayHomeAsUpEnabled(true);
+//          setProgressBarIndeterminateVisibility(true);
+//          recyclerView = (RecyclerView)findViewById(R.id.recycler_view);
+//          recyclerView.setLayoutManager(new LinearLayoutManager(this));
+//          recyclerView.setItemAnimator(new DefaultItemAnimator());
+//
+//
+//          //get already retrieved feed and reuse if it is there
+//          content = (ArrayList<Content>)getLastNonConfigurationInstance();
+//          if(content == null)
+//          {
+//              //no previous data, so database must be read
+//              readDB();
+//          }
+//          else
+//          {
+//              //reuse existing feed data
+//              adapter = new BookmarkRecyclerViewAdapter(content, R.layout.card_row, this);
+//              recyclerView.setAdapter(adapter);
+//              ItemTouchHelper.Callback callback = new SimpleItemTouchHelperCallback(adapter);
+//              itemTouchHelper = new ItemTouchHelper(callback);
+//              itemTouchHelper.attachToRecyclerView(recyclerView);
+//              setProgressBarIndeterminateVisibility(false);
+//
+//          }
       }
       
 
@@ -128,11 +144,11 @@ public class ViewBookmarksActivity extends Activity implements OnStartDragListen
      * For OnStartDragListener
      * @param viewHolder The holder of the view to drag.
      */
-    @Override
-    public void onStartDrag(RecyclerView.ViewHolder viewHolder)
-    {
-        itemTouchHelper.startDrag(viewHolder);
-    }
+//    @Override
+//    public void onStartDrag(RecyclerView.ViewHolder viewHolder)
+//    {
+//        itemTouchHelper.startDrag(viewHolder);
+//    }
 
     @Override
     public boolean onCreateOptionsMenu(Menu menu)
@@ -146,74 +162,74 @@ public class ViewBookmarksActivity extends Activity implements OnStartDragListen
      /* (non-Javadoc)
      * @see android.app.Activity#onResume()
      */
-    @Override
-      protected void onResume()
-      {
-          super.onResume();
-          //if database state has changed, reload the display
-          if(content != null)
-          {
-              int dbCount = getDBCount();
-              
-              if(dbCount >  content.size())
-              {
-                  readDB();
-              }
-          }
-      }
+//    @Override
+//      protected void onResume()
+//      {
+//          super.onResume();
+//          //if database state has changed, reload the display
+//          if(content != null)
+//          {
+//              int dbCount = getDBCount();
+//
+//              if(dbCount >  content.size())
+//              {
+//                  readDB();
+//              }
+//          }
+//      }
       
-      /** Actions after list is loaded in View*/
-      protected void finishedLoadingList() 
-      {
-          recyclerView.setAdapter(adapter);
-          ItemTouchHelper.Callback callback = new SimpleItemTouchHelperCallback(adapter);
-          itemTouchHelper = new ItemTouchHelper(callback);
-          itemTouchHelper.attachToRecyclerView(recyclerView);
-          setProgressBarIndeterminateVisibility(false);
-      }
-      
-      /** reads feed in a separate thread.  Starts progress dialog*/
-      private void readDB()
-      {
-          Thread loadFavsThread = new Thread() 
-          {
-            public void run() 
-            {
-
-                String order = "fav_icon DESC, fav_title ASC";
-                
-                content = DBUtils.readCursorIntoList(getContentResolver().query(Bookmarks.CONTENT_URI, null, null, null, order));
-                
-               Collections.sort(content);
-                
-                fillData(content);
-                handler.post(finishedLoadingListTask);
-            }
-          };
-          loadFavsThread.start();
-          
-      }
+//      /** Actions after list is loaded in View*/
+//      protected void finishedLoadingList()
+//      {
+//          recyclerView.setAdapter(adapter);
+//          ItemTouchHelper.Callback callback = new SimpleItemTouchHelperCallback(adapter);
+//          itemTouchHelper = new ItemTouchHelper(callback);
+//          itemTouchHelper.attachToRecyclerView(recyclerView);
+//          setProgressBarIndeterminateVisibility(false);
+//      }
+//
+//      /** reads feed in a separate thread.  Starts progress dialog*/
+//      private void readDB()
+//      {
+//          Thread loadFavsThread = new Thread()
+//          {
+//            public void run()
+//            {
+//
+//                String order = "fav_icon DESC, fav_title ASC";
+//
+//                content = DBUtils.readCursorIntoList(getContentResolver().query(Bookmarks.CONTENT_URI, null, null, null, order));
+//
+//               Collections.sort(content);
+//
+//                fillData(content);
+//                handler.post(finishedLoadingListTask);
+//            }
+//          };
+//          loadFavsThread.start();
+//
+//      }
       /**
        * Loads feed data into adapter on initial reading of feed
        * @param contentList ArrayList<Content>
        */
-      private void fillData(ArrayList<Content> contentList) 
-      {
-          //Log.d("LensViewer", "fillData() called");
-          adapter = new BookmarkRecyclerViewAdapter(contentList, R.layout.card_row,this);
-      }
+//      private void fillData(ArrayList<Content> contentList)
+//      {
+//          //Log.d("LensViewer", "fillData() called");
+//          adapter = new BookmarkRecyclerViewAdapter(contentList, R.layout.card_row,this);
+//      }
       
       /**
        * Queries the database to get the number of favorites stored
      * @return int - the number of favorites items in the database
      */
-    private int getDBCount()
-    {
-          Cursor c = getContentResolver().query(Bookmarks.CONTENT_URI, null, null, null, null);
-          int count = c.getCount();
-          c.close();
-          return count;
-          
-      }
+//    private int getDBCount()
+//    {
+//          Cursor c = getContentResolver().query(Bookmarks.CONTENT_URI, null, null, null, null);
+//          int count = c.getCount();
+//          c.close();
+//          return count;
+//
+//      }
 
 }
